fix(feature-svc): fix deadlock risk, empty catch, and decimal precision
- CryptoPricingService: replace .Result with await after Task.WhenAll
- ServiceRegistration: add logging to empty catch block in JWT fallback
- DocumentService + DocumentDtos: double → decimal for percentage calc

+++ feature-service/src/FeatureService.Api/DTOs/DocumentDtos.cs
@@ -104,7 +104,7 @@ public record StorageQuotaDto(
     long UsedBytes,
     long MaxBytes,
     int DocumentCount,
-    double UsedPercentage
+    decimal UsedPercentage
 );
 
 /// <summary>
